% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XCos.R
\name{XCosScore}
\alias{XCosScore}
\title{XCos}
\usage{
XCosScore(refMatrix, query, topN = 500, permuteNum = 10000,
  pAdjMethod = "BH", mcCore = 1)
}
\arguments{
\item{refMatrix}{A matrix}

\item{query}{A numeric vector with names.}

\item{topN}{number of genes in top and bottom of reference gene list}

\item{permuteNum}{number of perturbation time for computing pValue}

\item{pAdjMethod}{method to use for computing adjudted-pValue}

\item{mcCore}{the number of core to use for parallel computing}
}
\description{
The implementation of XCos
}
\details{
`XCosScore()` returns a data.frame, each row of which contains
  score, pValue and adjusted-pValue for one sample in the refMatrix.
}
\examples{
set.seed(1234)
ref <- matrix(rnorm(1000), nrow = 10,
  dimnames = list(paste0("gene", 1:10), paste0("drug", 1:100)))
query <- rnorm(4)
names(query) <- c("gene1", "gene2", "gene9", "gene10")
XCosScore(refMatrix = ref, query = query, topN = 4)
}
\references{
"Cheng J et al. Genome medicine, 2014, 6(12): 95".
}
