% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XSum.R
\name{XSumScore}
\alias{XSumScore}
\title{XSum}
\usage{
XSumScore(refMatrix, queryUp, queryDown, topN = 500,
  permuteNum = 10000, pAdjMethod = "BH", mcCore = 1)
}
\arguments{
\item{refMatrix}{A matrix}

\item{queryUp, queryDown}{character vectors.}

\item{topN}{number of genes in top and bottom of reference gene list}

\item{permuteNum}{number of perturbation time for computing pValue}

\item{pAdjMethod}{method to use for computing adjudted-pValue}

\item{mcCore}{the number of core to use for parallel computing}
}
\description{
The implementation of XSum
}
\details{
`XSumScore()` returns a data.frame, each row of which contains score,
  pValue and adjusted-pValue for one sample in the refMatrix.
}
\examples{
set.seed(1234)
ref <- matrix(rnorm(1000), nrow = 10,
  dimnames = list(paste0("gene", 1:10), paste0("drug", 1:100)))
Up <- c("gene1", "gene2")
Down <- c("gene9", "gene10")
XSumScore(refMatrix = ref, queryUp = Up, queryDown = Down, topN = 4)
}
\references{
"Cheng J et al. Genome medicine, 2014, 6(12): 95".
}
