% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZhangScore.R
\name{ZhangScore}
\alias{ZhangScore}
\title{ZhangScore}
\usage{
ZhangScore(refMatrix, queryUp, queryDown, permuteNum = 10000,
  pAdjMethod = "BH", mcCore = 1)
}
\arguments{
\item{refMatrix}{A matrix}

\item{queryUp, queryDown}{character vectors.}

\item{permuteNum}{number of perturbation time for computing pValue}

\item{pAdjMethod}{method to use for computing adjudted-pValue}

\item{mcCore}{the number of core to use for parallel computing}
}
\description{
The implementation of ZhangScore
}
\details{
`ZhangScore()` returns a data.frame, each row of which contains score,
pValue and adjusted-pValue for one sample in the refMatrix.
}
\examples{
set.seed(1234)
ref <- matrix(rnorm(1000), nrow = 10,
  dimnames = list(paste0("gene", 1:10), paste0("drug", 1:100)))
Up <- c("gene1", "gene2")
Down <- c("gene9", "gene10")
ZhangScore(refMatrix = ref, queryUp = Up, queryDown = Down)
}
\references{
"Zhang S D et al. BMC bioinformatics, 2008, 9(1): 258".
}
